+++ packages/types/src/primitive/I256.ts
@@ -8,4 +8,7 @@ import { Int } from '../codec/Int';
  * @description
  * A 256-bit signed integer
  */
-export class i256 extends Int.with(256) {}
+export class i256 extends Int.with(256) {
+  // NOTE without this, we cannot properly determine extensions
+  public readonly __IntType = 'u256';
+}