+++ packages/types/src/primitive/U16.ts
@@ -8,4 +8,7 @@ import { UInt } from '../codec/UInt';
  * @description
  * A 16-bit unsigned integer
  */
-export class u16 extends UInt.with(16) {}
+export class u16 extends UInt.with(16) {
+  // NOTE without this, we cannot properly determine extensions
+  public readonly __UIntType = 'u16';
+}