+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes:
 
 - Loose pedantic check not to fail on all-empty values
 - Extend `createType` auto-TS decoration for `BreeMap`, `BTreeSet`, `HashMap`
+- Convert `TypeDefInfo` to TS union type from enum
 
 
 ## 5.4.1 Aug 9, 2021